# task Users/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample express server.",
    "version": "1.0.0",
    "title": "Swagger Doc",
    "contact": {
      "email": "bratskov.alexander@gmail.com"
    },
    "license": {
      "name": "ISC"
    }
  },
  "host": "localhost:3000",
...
  "schemes": ["http"],
...
  "paths": {
...
    "/users/{userId}": {
      "put": {
        "tags": ["users"],
        "summary": "Update user by ID",
        "description": "return update user",
        "operationId": "updateUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User object that needs to be updated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/user"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete user",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteUser",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful"
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "user": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "cityID": {
          "type": "string"
      }
    }
  }
}
}